# ddl: Speed up add index phase