Handle issue #2257 
- Use `BatchGet` instead of `Get`

- Remove the judgment of `Exist`

Using 3 TiKV, 1 PD, 1 TiDB , the servers of TiKV and PD on one computer, the server of TiDB on another computer.
The lease of TiDB is 1s.
The number of records in this table is 336253, and the table's structure is as follows:
+------------------+------------------+------+------+---------+-------+
| Field            | Type             | Null | Key  | Default | Extra |
+------------------+------------------+------+------+---------+-------+
| a           | varchar(64)      | YES  |      |         |       |
| b           | bigint(20)       | NO   |      | 0       |       |
| c           | datetime         | YES  |      | NULL    |       |
| d           | int(11) UNSIGNED | NO   |      | 0       |       |
| e           | int(11) UNSIGNED | NO   |      | 0       |       |
| f            | int(11) UNSIGNED | NO   |      | 0       |       |
| g           | varchar(28)      | NO   |      |         |       |
| h           | varchar(28)      | NO   |      |         |       |
+------------------+------------------+------+------+---------+-------+

tidb-before> alter table t add index id (b);
Query OK, 0 rows affected (6 min 2.55 sec)
tidb-before> alter table t add index ids (e);
Query OK, 0 rows affected (6 min 9.32 sec)

tidb-current> alter table t add index id (b);
Query OK, 0 rows affected (1 min 6.78 sec)
tidb-current> mysql> alter table t add index ids (e);
Query OK, 0 rows affected (1 min 3.15 sec)