@disksing 2016-12-16
It will be better to break `fetchRowColVals` into following steps:
1) Encode `handles []uint64` to `rowkeys []kv.Key`.
2) `BatchGet(rowKeys)`, you will get `pairMap map[kv.Key][]byte`.
3) Iterate `rowKeys` to build `idxRecords []*indexRecord`, just ignore keys that are absent, you don't  need the `notExist` flag.